🧹 extract io writer to iox package
